+++ cmd/databricks-claude/token.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/IceRhymers/databricks-agents/internal/core/cli"
-	"github.com/IceRhymers/databricks-agents/pkg/tokencache"
+	"github.com/IceRhymers/databricks-agents/internal/core/tokencache"
 )
 
 // resolveDatabricksCLI delegates to pkg/cli for CLI binary resolution.
