+++ internal/core/tokencache/AGENTS.md
